+++ packages/react-magma-dom/src/components/Select/MultiSelect.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { MultiSelectProps } from '.';
+import { instanceOfDefaultItemObject, MultiSelectProps } from '.';
 import { useMultipleSelection, useSelect } from 'downshift';
 import { CloseIcon } from 'react-magma-icons';
 import { ItemsList } from './ItemsList';
@@ -9,6 +9,8 @@ import { IconWrapper, SelectedItemButton, SelectText } from './shared';
 
 import { ThemeContext } from '../../theme/ThemeContext';
 import { I18nContext } from '../../i18n';
+import { ButtonSize, ButtonVariant } from '../Button';
+import { defaultComponents } from './components';
 
 export function MultiSelect<T>(props: MultiSelectProps<T>) {
   const {
@@ -42,6 +44,7 @@ export function MultiSelect<T>(props: MultiSelectProps<T>) {
     placeholder,
     setFloating,
     setReference,
+    isClearable,
   } = props;
 
   function checkSelectedItemValidity(itemToCheck: T) {
@@ -57,6 +60,7 @@ export function MultiSelect<T>(props: MultiSelectProps<T>) {
     removeSelectedItem,
     selectedItems,
     setActiveIndex,
+    reset,
   } = useMultipleSelection<T>({
     ...props,
     ...(props.initialSelectedItems && {
@@ -162,6 +166,44 @@ export function MultiSelect<T>(props: MultiSelectProps<T>) {
     disabled: disabled,
   });
 
+  const { ClearIndicator } = defaultComponents<T>({
+    ...customComponents,
+  });
+
+  function itemsArrayToString(itemsArray: any[]) {
+    let allItems = [];
+    itemsArray.map(item => {
+      if (typeof item === 'string') {
+        allItems.push(item);
+      } else if (instanceOfDefaultItemObject(item)) {
+        allItems.push(item.label);
+      }
+    });
+
+    return allItems.join(', ');
+  }
+
+  const toggleButtonRef = React.useRef<HTMLButtonElement>();
+
+  const clearIndicatori18n =
+    selectedItems.length > 1
+      ? i18n.select.multi.clearIndicatorAriaLabel
+      : i18n.select.clearIndicatorAriaLabel;
+
+  const clearIndicatorAriaLabel = clearIndicatori18n
+    .replace(/\{labelText\}/g, labelText)
+    .replace(/\{selectedItem\}/g, itemsArrayToString(selectedItems));
+
+  function defaultHandleClearIndicatorClick(event: React.SyntheticEvent) {
+    event.stopPropagation();
+
+    if (toggleButtonRef.current) {
+      toggleButtonRef.current.focus();
+    }
+
+    reset();
+  }
+
   return (
     <SelectContainer
       additionalContent={additionalContent}
@@ -220,6 +262,25 @@ export function MultiSelect<T>(props: MultiSelectProps<T>) {
           <SelectText>{i18n.multiSelect.placeholder}</SelectText>
         )}
       </SelectTriggerButton>
+
+      {isClearable && selectedItems?.length > 0 && (
+        <ClearIndicator
+          aria-label={clearIndicatorAriaLabel}
+          icon={<CloseIcon size={theme.iconSizes.xSmall} />}
+          isInverse={isInverse}
+          onClick={defaultHandleClearIndicatorClick}
+          size={ButtonSize.small}
+          variant={ButtonVariant.link}
+          disabled={disabled}
+          testId="clearIndicator"
+          style={{
+            position: 'absolute',
+            right: '3.25em',
+            top: '50%',
+            transform: 'translateY(-50%)'
+          }}
+        />
+      )}
       <ItemsList
         customComponents={customComponents}
         floatingElementStyles={floatingElementStyles}