+++ packages/react-magma-dom/src/components/Select/SelectTriggerButton.tsx
@@ -29,6 +29,7 @@ const ChildrenContainer = styled.div`
   flex-grow: 1;
   flex-wrap: wrap;
   min-height: ${props => props.theme.spaceScale.spacing09};
+  min-width: 0%;
 `;
 
 interface SelectTriggerButtonInterface<T> {