+++ patterns/header/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 12.0.0-next.1
+
+### Patch Changes
+
+- Updated dependencies [c815b89ac]
+  - react-magma-dom@4.7.0-next.1
+
 ## 11.0.1-next.0
 
 ### Patch Changes